refactor(cache): restructure GlobalCache to Fluent Facade with Ops interfaces and capacity guards
- Refactor GlobalCache from 2600+ line monolith to 143-line static facade
- Introduce 15 typed Ops interfaces (Value/Struct/Field/Collection/Ranking/Lock/Bitmap/HyperLog/Geo/BoundedQueue/BoundedStack/Key/Script/Limiter/Notification)
- Extract CacheInfrastructure from KeyOps for framework-internal capabilities (L2 switch, type registration)
- Add Set capacity protection: SCARD guard (5K WARN / 10K reject) in CollectionOps
- Add ZSet capacity protection: ZCARD guard (5K WARN / 10K reject) in RankingOps
- Change BoundedQueue/BoundedStack destroy() from DEL to UNLINK for async cleanup
- Rewrite README: align big-key thresholds with RedisPerfGuard defaults and encoding thresholds
- Remove Redis Stream module (moved to separate module atlas-richie-component-redis-streammq)
- Adapt all downstream consumers: gateway, mfa, messaging, storage, statemachine, logging, mqtt, i18n, sample-cache

+++ atlas-richie-component/atlas-richie-component-storage/atlas-richie-component-storage-minio/src/main/java/com/richie/component/storage/config/MinioAutoConfiguration.java
@@ -89,7 +89,7 @@ public class MinioAutoConfiguration {
         try {
             client.putObject(
                             PutObjectArgs.builder().bucket(bucket).object(key)
-                                    .stream(new ByteArrayInputStream(bytes), bytes.length, -1)
+                                    .stream(new ByteArrayInputStream(bytes), (long) bytes.length, -1L)
                                     .build())
                     .get(30, TimeUnit.SECONDS);
             try (var is = client.getObject(GetObjectArgs.builder().bucket(bucket).object(key).build())
